# Conflicts:
#	src/Arch/Mos6502/Mos6502ProcessorArchitecture.cs
#	src/Core/Core.csproj
#	src/Core/ImageWriter.cs
#	src/Core/MemoryArea.cs
#	src/Core/StructureReader.cs
#	src/Drivers/CmdLine/CmdLine.csproj
#	src/Reko-decompiler.sln
#	src/UnitTests/ImageLoaders/MzExe/PeImageLoaderTests.cs
#	src/UnitTests/UnitTests.csproj

+++ src/UnitTests/ImageLoaders/Hunk/HunkMaker.cs
@@ -52,7 +52,7 @@ namespace Reko.UnitTests.ImageLoaders.Hunk
             {
                 dispatcher[o.GetType()](o, w);
             }
-            var bytes = w.Bytes.Take(w.Position).ToArray();
+            var bytes = w.Bytes.Take((int)w.Position).ToArray();
             return bytes;
         }
